fix: restructure sdk client files
