--- packages/sdk/src/client/actions/index.ts
@@ -1,2 +1,0 @@
-export * from "./account.js";
-export * from "./session.js";
