--- packages/sdk/src/client/decorators/index.ts
@@ -1,2 +1,0 @@
-export * from "./session.js";
-export * from "./session_wallet.js";
