+++ packages/sdk/src/client/index.ts
@@ -1,2 +1,2 @@
-export { deployAccount, fetchAccount } from "./actions/account.js";
-export * from "./clients/session.js";
+export { deployAccount, fetchAccount } from "./passkey/actions/account.js";
+export * from "./session/client.js";
