+++ packages/sdk/src/client/passkey/account.ts
@@ -3,29 +3,21 @@ import { type CustomSource, type Hash, hashMessage, hashTypedData, type Hex, typ
 import { toAccount } from "viem/accounts";
 import { serializeTransaction, type ZksyncTransactionSerializableEIP712 } from "viem/zksync";
 
-import { getEip712Domain } from "./utils/getEip712Domain.js";
+import { getEip712Domain } from "../utils/getEip712Domain.js";
 
-export type ToSmartAccountParameters = {
+export type ToPasskeyAccountParameters = {
   /** Address of the deployed Account's Contract implementation. */
   address: Address;
-  /** Function to sign a hash. */
   sign: (parameters: { hash: Hash }) => Promise<Hex>;
 };
 
-export type ZksyncSmartAccount = LocalAccount<"smartAccountZksync"> & {
+export type PasskeyAccount = LocalAccount<"ssoPasskeyAccount"> & {
   sign: NonNullable<CustomSource["sign"]>;
 };
 
-type ErrorType<name extends string = "Error"> = Error & { name: name };
-export type ToSmartAccountErrorType = ErrorType;
-
-/**
- * Creates a [ZKsync Smart Account](https://docs.zksync.io/build/developer-reference/account-abstraction/building-smart-accounts)
- * from a Contract Address and a custom sign function.
- */
-export function toSmartAccount(
-  parameters: ToSmartAccountParameters,
-): ZksyncSmartAccount {
+export function toPasskeyAccount(
+  parameters: ToPasskeyAccountParameters,
+): PasskeyAccount {
   const { address, sign } = parameters;
 
   const account = toAccount({
@@ -62,6 +54,6 @@ export function toSmartAccount(
 
   return {
     ...account,
-    source: "smartAccountZksync",
-  } as ZksyncSmartAccount;
+    source: "ssoPasskeyAccount",
+  } as PasskeyAccount;
 }
