+++ packages/sdk/src/client/passkey/index.ts
@@ -1,3 +1,4 @@
+export * from "./account.js";
 export * from "./actions/passkey.js";
-export * from "./clients/passkey.js";
+export * from "./client.js";
 export * from "./decorators/passkey.js";
